Add alternate #-style commenting
This commit adds a 'style' option which allows the caller to choose from
a list of pre-defined styles for comments to be parsed from the target
file. The default for this option is 'C-style', which selects the same
behavior as this extension has provided to date.

Additionally, this adds a 'hash' style option, which extracts comments
of the following format, intended for Ansible (or YAML, generally)
files.

  #:
  # This is a sample comment which would be selected with the 'hash'
  # style option.
  #.
  - name: An Ansible Task
    ...