Modify whitespace handling to reflect \code placement
Previously, when the \code directive was supplied, the subsequently
included lines were added verbatim, plus one leading space. This change
adjusts the handling of the indentation stack so that the included code
is always indented to match the placement of the \code directive (plus
any other leading whitespace the line includes).