Add alternate #-style commenting (#6)
* Add alternate #-style commenting

This commit adds a 'style' option which allows the caller to choose from
a list of pre-defined styles for comments to be parsed from the target
file. The default for this option is 'C-style', which selects the same
behavior as this extension has provided to date.

Additionally, this adds a 'hash' style option, which extracts comments
of the following format, intended for Ansible (or YAML, generally)
files.

  #:
  # This is a sample comment which would be selected with the 'hash'
  # style option.
  #.
  - name: An Ansible Task
    ...

* Permit leading whitespace on hash-style comments

* Correctly handle blank hash comment lines

* Modify whitespace handling to reflect \code placement

Previously, when the \code directive was supplied, the subsequently
included lines were added verbatim, plus one leading space. This change
adjusts the handling of the indentation stack so that the included code
is always indented to match the placement of the \code directive (plus
any other leading whitespace the line includes).

* Switch to regex for \code tag detection

This also permits a better method for whitespace detection, and fixes a
bug in which the pre-comment whitespace was detected instead of the
indent into the comment.

* Add \codepara marker for easier markup